move tests to crate
